nginx{,-module}.eclass: do not fail if econf_ngx --help is called
NGINX's ./configure return 1 if "--help" is passed. edo does not like it
and thinks the command failed. The previous code was incorrect in how it
handled the "--help" case.

edo actually never failed (since it was always called with nonfatal,
even when no "--help" was passed), only showed the edo's "Failed to run
command" message.

This commit fixes the --help and all other cases altogether, making
econf_ngx fail when necessary and shut when not.

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/42895
Closes: https://github.com/gentoo/gentoo/pull/42895
Signed-off-by: Sam James <sam@gentoo.org>