sys-apps/musl-locales: Create env file
With vimproved's work to add musl-locale to @system, it also make sense to streamline the install process for users so the env file is created during the install.

I've choosen to -r1 to make it easier to track issues that might occur, even if it likely is overkill for env file.

Signed-off-by: Ian Jordan <immoloism@gmail.com>