media-plugins/ams-lv2: enable py3.12, py3.13, py3.14
The package performs crimes against technology: namely, it builds with
waf. This is the only reason it needs python at all.

This necessitates replacing bundled waf with a more modern vesion, since
the bundled version was already crudely patched to support py3.11
instead of updating waf, and it immediately broke again for 3.12. The
current replacement waf is used elsewhere in ::gentoo, and is tested to
install the same using any python from 3.11-3.14.

Closes: https://bugs.gentoo.org/929650
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>