media-plugins/ams-lv2: cleanup dependencies a bit
gtkmm:2.4 is a C++ wrapper around gtk+:2, and installing the former
implies the latter. It's redundant to specify both, and also incorrect
since the latter is never actually used directly. qa-vdb even points out
it isn't used.

Same logic applies to cairo vs cairomm, but this package only listed the
(unused) C part. This is fine, because gtkmm noisily adds it itself, but
qa-vdb does point out cairomm is used.

No revbump because this makes no difference at either build or install
time. The deps are there regardless, but lint tools are a bit happier is
all.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>