dev-vcs/git-cliff: new package, add 2.9.1
dev-vcs/git-cliff: new package, add 2.9.1
This PR aims to get `git-cliff` inside the official Gentoo portage-tree.

`git-cliff` is a Git extension for automated changelog creation.

As such a helper might be a benefit for staying close to conventional-commits
as well as keeping an organized Git structure. This extension might help or
support one or another to get rid of annoying changelog issues.

This PR adds the following ***ebuild***:
* `git-cliff-2.9.1.ebuild`: latest "tagged" version

And the following ***patches***:
* `git-cliff-2.9.1-disable_repo_tests.patch`:
  disables git-repository related tests (not available inside sandbox)
* `git-cliff-2.9.1-silence_run_os_command_test.patch`:
  disables failing OS-command test (not available inside sandbox)

Adding myself as the maintainer of this ebuild.

Signed-off-by: Armas Spann <zappel@simple-co.de>
Part-of: https://github.com/gentoo/gentoo/pull/42924
Closes: https://github.com/gentoo/gentoo/pull/42924
Signed-off-by: Maciej Barć <xgqt@gentoo.org>