net-analyzer/wapiti: update fix absolutely trash test restriction
> Many tests require execution of local test php server

is surely not true, it simply needs dev-lang/php installed. So, there's
no earthly reason to restrict it.

Once we try running tests, there is a cascade of failures. The biggest
one is the source tarball doesn't come with them -- perhaps this would
be noticed if we tried running them, eh?

There are a *bunch* of other missing test deps too.

And while we are here -- remove the need for pytest-cov via sed on
"addopts". There's never a good reason to depend on linting plugins just
because pytest.ini (or equivalent) enables it by default in CI.

Bug: https://bugs.gentoo.org/952156
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>