sys-devel/gcc: keyword 15.1.1_p20250705-r1
Propagate the fix from cdcd99a6abad0a6c524d2edb73aecdfc5e62244c. Was ready
to keyword a new snapshot from the branch already but was waiting for this
and/or the Ada bug (PR120440) but the latter isn't fixed yet and we shouldn't
wait to propagate this Qt fix.

Closes: https://bugs.gentoo.org/956308
Signed-off-by: Sam James <sam@gentoo.org>