www-nginx/ngx-vod: fix unprototyped func declaration causing build errors on new gcc's
Before version 15, GCC had -std=gnu17 as the default. GCC 15 bumped the
default to -std=gnu23. One of the changes in C23 is the interpretation
of function declaration without parameters (unprototyped declaration).

In C17 and before, the unprototyped declaration specified no information
on the number and types of arguments. With GCC 15, the declaration of
the form
    int myfunc()
is treated the same as
    int myfunc(void)
i.e. the function takes no arguments.

This change broke compilation of the module which relied on older
interpretation. The broken function, having the unprototyped
declaration, actually takes one argument of type ngx_cycle_t *.

This commit corrects the function declaration and definition, explicitly
specifying the arguments.

Closes: https://bugs.gentoo.org/959586
Closes: https://bugs.gentoo.org/959690
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>