www-nginx/ngx-brotli: revbump 20231109 to r1, use system libbrotlienc
This imports a patch[1] from upstream GitHub PR. The patch links to the
system Brotli lib using pkgconfig.

Since the system Brotli lib is used in favour of the bundled Brotli lib,
the commit drops bundled Brotli from the package.

Many thanks to Mikel Olasagasti Uranga for writing the patch[1].

[1]: https://github.com/google/ngx_brotli/pull/172

Closes: https://bugs.gentoo.org/959780
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>