gui-apps/foot: remove likely redundant install ops
These install steps may be unnecessary as foot does not seem to actually
unconditionally install the three files and bypass scdoc (which looks
like is mistakenly referred to as dodoc in the ebuild files?). I may be
wrong on their necessity though, so please revert or otherwise modify
this commit and what it does if that is the case.

Signed-off-by: asdermy <220292596+asdermy@users.noreply.github.com>