sys-kernel/gentoo-kernel-bin: increase ebuild symmetry
Instead of setting KERNEL_EFI_ZBOOT, we add a check for it to all
ebuilds. We also set KERNEL_IUSE_MODULES_SIGN=1 here because the
kernel-install eclass enables enforcing of module signatures in the
test phase conditional to this variable. This is temporary until we
finish symmetrising the older kernel series with the newer ones, then
we can just drop this KERNEL_IUSE_MODULES_SIGN guard from the eclasses
entirely.

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>