x11-terms/kitty-shell-integration: install only auto scripts in live
See bug #960075 for details, but (as far as I can tell) remotely the
auto-loaded scripts (e.g. bashrc.d) is all that should be needed from
the shell-integration directory, the rest can be kept in the main
kitty package.

Unfortunately no such location for zsh, but still install the scripts
somewhere in case users need it but do not want to install kitty.
Users can opt to resort to `kitten ssh` instead though.

This also drops completions which did not make much sense if kitty is
not also installed, besides these were incomplete and it'll be handled
hopeufully properly in the kitty package instead.

Bug: https://bugs.gentoo.org/960075
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>