x11-terms/kitty: install shell completion files in live
Some of this was already done by x11-terms/kitty-shell-integration but
there was little reason to have that package install these (see
bug #960075) and the pre-generated files it installed were missing
things (most notably it did not have a pre-gen file for bash).

Note that for 9999 this will cause conflicts unless update
kitty-shell-integration first (normally these are always supposed
to be updated together and in order either way, albeit for 9999
that is not enforced). Otherwise blockers are not needed given
kitty pins to exact versions forcing the update first.

Bug: https://bugs.gentoo.org/960075
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>