x11-terms/kitty: stop forcing no-rc in live
As noted in bug #960075, this was formerly passed to prevent kitty
from modifying e.g. .bashrc out-of-the-box when it wasn't needed.
However it no longer modify files as far as I can tell see, and
this helps zsh which lacks a meant to auto-load the kitty
integration on Gentoo at the moment. Also won't load twice if files
were already loaded making it harmless for non-zsh, and having two
methods ensure it will be loaded in more situations (remote,
sub-shells, tmux, custom setups, etc...).

Bug: https://bugs.gentoo.org/960075
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>