gui-apps/foot: add `doc` USE flag to latest ver
This USE flag corresponds directly with foot's `docs` compile-time
option. The flag is set to be enabled by default as it is
unconditionally enabled for previous versions, and users upgrading their
installation of foot to this version when it is marked as stable in the
future may unexpectedly be unable to find the documentation files they
assume will still be there.

Signed-off-by: asdermy <220292596+asdermy@users.noreply.github.com>