netperf: Inherit the autotools eclass
The ebuild for version 2.7.0 (netperf-2.7.0-r5.ebuild) doesn't include
'make' in its dependency graph, which may cause it to fail to emerge
with 'make: command not found'.
There is also netperf-2.7.0_p20210121.ebuild, and quickly diff'ing the
two reveals that the latter is inheriting autotools, pulling in make
to the dependency graph.
Add this inherit in the 2.7.0 version.

Bug: 960348
Signed-off-by: Sergio Gomez <sergio.g.delreal@gmail.com>