net-misc/sunshine: Build and statically link Boost 1.87 for now
Unfortunately, Sunshine relies heavily on the v1 process API that was dropped in
Boost 1.88. I tried to make the necessary changes, but v2 is very different and
missing features present in v1. I'll leave this mess to upstream. In the
meantime, build and statically link 1.87. This is less than ideal but better
than holding everyone back. It raises a couple of CMake compatibility warnings,
but upstream will hopefully bump Boost before this becomes an issue.

Closes: https://bugs.gentoo.org/955560
Signed-off-by: James Le Cuirot <chewi@gentoo.org>