games-roguelike/dwarf-fortress: Version bump to 51.13, now binary-only
Upstream still ships the source code for the graphical component, but it
has frequently been unusable or has at least required heavy patching. We
are no longer interested in trying to fix code that is so blatantly
broken. There is little need to build this from source when the rest is
proprietary anyway.

The game writes its save data among the rest of the game files. This was
previously worked around using copies and symlinks, but this was messy,
and went against upstream's advice of not extracting newer versions over
the top of older versions. We now set up an overlay with bubblewrap
instead.

The user data directory has moved from ~/.dwarf-fortress to
~/.local/share/dwarf-fortress. No old save data is migrated because
version 50.01 broke compatibility anyway.

32-bit x86 is no longer supported by upstream.

Closes: https://bugs.gentoo.org/702918
Closes: https://bugs.gentoo.org/739328
Closes: https://bugs.gentoo.org/831875
Closes: https://bugs.gentoo.org/906145
Closes: https://bugs.gentoo.org/937581
Closes: https://bugs.gentoo.org/952854
Signed-off-by: James Le Cuirot <chewi@gentoo.org>