x11-misc/lightdm: remove libxml2 dep entirely as unused
Originally added in gentoo-historical-2.git commit
23500cafc46bebf85205eac51a0ead46a2d46431 titled "Re-introduce lightdm
package." as of September 2011. At the time, the source code contained
few greps for "xml, all of which essentially boiled down to "gtk-doc",
"the file format for an svg icon", and "g_dbus_node_info_new_for_xml".
No use of libxml2 at all.

I'm baffled as to why it was ever added. libxclavier adds transitive
cflags-only for libxml2 via Requires.private, and isn't a true
dependency. yelp-tools has a runtime dep on `xmllint` and checks for it
via yelp.m4 -- but originally no autoreconf was done, hence no bdep on
yelp-tools. Still, that would be a *build* time dep only. What gives?

Demote the dependency to a bdep.

Bug: https://bugs.gentoo.org/365311
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>