net-misc/sunshine: Fix building with Boost 1.87 installed
We want to build an internal Boost even if 1.87 is installed.
Unfortunately, CMake doesn't handle this well. It finds Boost itself but
fails to find all the components because they don't have static
libraries. Even if Sunshine were to explicitly check for the presence of
each component, it would then fail because the Boost::headers target
cannot be redefined. I tried to fix this properly, but it seems
impossible. I thought that CMAKE_DISABLE_FIND_PACKAGE would prevent
Boost from working at all, but it somehow did the trick.

Closes: https://bugs.gentoo.org/960465
Signed-off-by: James Le Cuirot <chewi@gentoo.org>