net-dns/dnsdist: add 2.0.0
This is a major release with lots of fixes and new features, but
just as stable and reliable as the 1.9.x line; upgrading should be
seamless as existing configurations continue to work without problems.
The new YAML-based configuration is optional since it requires Rust.

In theory the ebuild could have migrated to meson, but to reduce the
risks of the update we continue to use autools for now.

Signed-off-by: Holger Hoffstätte <holger@applied-asynchrony.com>