net-misc/yt-dlp: add 2025.07.21
Note that the test_utils.py::TestUtil::test_extract_attributes test is
known failing specifically with python-3.14.0_beta4, but according to
upstream should work again with the next python release.

Afaik does not cause notable runtime issues, so not going to remove
compat, skip the test, or keep an ebuild note over this if it'll pass
again soon.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>