dev-db/mydumper: correct PCRE dependencies
Upstream switched to a PCRE2 in version v0.18.1.1, see commit [1,2].
Therefore, this change corrects dependencies to a library which is
actually used in runtime. However, upstream removed remaining unused
pcre.h includes later in version v0.18.2.11, see commit [3], which makes
PCRE still implicitly required in buildtime for this stable version due
to lingering headers in the source.

I've decided to address the buildtime issue by adding libpcre as a
DEPEND, because I prefer not to modify the code of stable versions,
especially since the patch needed is quite large. Plus, users already
need both libpcre and libpcre2 installed to avoid buildtime failures
anyway. Moreover, the fixed version is already in the tree. We just have
to wait approprite time to stabilize it and then we can clean this one.

[1] https://github.com/mydumper/mydumper/commit/9de78071ae73b40bb972fcbdb328ad20fa562241
[2] https://github.com/mydumper/mydumper/commit/c098a1407e0f06cb447f0c730bf9f0badc8706e3
[3] https://github.com/mydumper/mydumper/commit/9e10f0083ea6419d17265a0bb82453be0cadb633

Signed-off-by: Petr Vaněk <arkamar@gentoo.org>