sci-mathematics/polymake: add 4.14
Adds "+bliss" to IUSE to finally fix the pkgcheck warning. One of
bliss/nauty are required, and bliss is the lighter-weight of the two, so
it makes a better default. Our nauty package also tends to lag a bit
because we use several of Debian's patches that need to be rerolled for
each release.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>