kernel-build.eclass: Force GNU tar for kernel builds
Override the tar executable used by kernel Makefiles to force GNU tar.
This is necessary since some of the scripts use options specific
to GNU tar, and therefore fail with other choices for
app-alternatives/tar.  That said, this override works only with the most
recent version of sys-kernel/gentoo-kernel.  It doesn't cover all
upstream uses of tar, but it does cover these that fail with other tar
implementations.

Closes: https://bugs.gentoo.org/951102
Signed-off-by: Michał Górny <mgorny@gentoo.org>