x11-drivers/nvidia-drivers: drop vuln 550.163.01-r1, end of 550.x
It "seems" NVIDIA has ended support for this branch and it may be
affected by the vulnerability in bug #960647. Users should either
upgrade to 570.x or, if have problems with that branch, downgrade
to the latest 535.x which appears still supported and has fixed the
vulnerability (unknown for how much longer 535.x will be supported
though, ideally try to get >=570.x to work for your setup).

Bug: https://bugs.gentoo.org/960647
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>