dist-kernel-utils.eclass: reinstall_initramfs() zboot support
If the config still exists, check it for CONFIG_EFI_ZBOOT, if it is
then the name of the kernel image will be different. We already
introduced the KERNEL_EFI_ZBOOT=1 variable to instruct the
dist-kernel_get_image_path() function to look for this different name.

We however forgot to add some logic to set it for the
reinstall_initramfs() path which is called via linux-mod-r1.eclass
by for example sys-fs/zfs-kmod[dist-kernel,initramfs]. This ebuild
has no information on whether the target kernel was built with ZBOOT
enabled or not. So we add now some logic to the eclass to check this.

Closes: https://bugs.gentoo.org/960603
Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>