qmake-utils.eclass: add "-spec ..." for qmake6
If qt6 is built with clang, QMAKE_SPEC defaults to linux-clang.
However, if we try to build, for example, "app-benchmarks/i7z" with gcc,
if will fail with:

> g++: error: unrecognized command-line option '-fno-direct-access-external-data'

This suggests that setting QMAKE_CC / QMAKE_CXX / ... alone is not
sufficient for qmake to select the correct compiler. See the test
case below:

> # cat a.pro
> message(QMAKE_CC is $$QMAKE_CC)
> clang {
>         message("clang is true")
> } else {
>         message("clang is false")
> }
> # /usr/lib64/qt6/bin/qmake a.pro
> Project MESSAGE: QMAKE_CC is clang
> Project MESSAGE: clang is true
> # /usr/lib64/qt6/bin/qmake a.pro QMAKE_CC=gcc
> Project MESSAGE: QMAKE_CC is gcc
> Project MESSAGE: clang is true

After adding "-spec linux-g++",

> # /usr/lib64/qt6/bin/qmake a.pro QMAKE_CC=gcc -spec linux-g++
> Project MESSAGE: QMAKE_CC is gcc
> Project MESSAGE: clang is false

Please note: for now, we only handle Linux with GCC or Clang.

Closes: https://bugs.gentoo.org/960633
Signed-off-by: Z. Liu <zhixu.liu@gmail.com>