greadme.eclass: drop "file exists" check on append
In hindsight, this check is too restrictive. The check makes it
impossible to have a completely optional README.gentoo created. For
example, a README.gentoo where all the content is conditional on USE
flags.

To address this, we drop the "file exists" check on append. The rest of
the eclass already correctly handles cases where README.gentoo might not
exist.

Signed-off-by: Florian Schmaus <flow@gentoo.org>