app-accessibility/espeakup: Add module autoloading
Adding the speakup_soft module to be autoloaded when installing this package as it will fail launch the service without it enabled.

This will greatly improve accessibility under Gentoo including the offerings we support in the live media.

Bumped to EAPI8 as a freebie.

Signed-off-by: Ian Jordan <immoloism@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/43178
Closes: https://github.com/gentoo/gentoo/pull/43178
Signed-off-by: Sam James <sam@gentoo.org>