app-accessibility/espeakup: Add module autoloading
Adding the speakup_soft module to be autoloaded when installing this package as it will fail launch the service without it enabled.

This will greatly improve accessibility under Gentoo including the offerings we support in the live media.

revbumping to help keep track of new issues caused by this.

Bumped to EAPI8 as a freebie.

Signed-off-by: Ian Jordan <immoloism@gmail.com>