net-misc/ifenslave: explain why a variable is "NOT an error"
Explain the mysterious MY_PN comment which is "NOT an error" (it indeed
isn't, but why isn't it?) after some research: Debian's 2.x is a rewrite,
not using C, but ifupdown hooks instead. We don't want that.

(That said, ifenslave seems obsolete to me, netifrc supports bonding
for a long time..)

Signed-off-by: Sam James <sam@gentoo.org>