net-misc/ifenslave: EAPI 8, update Debian sources from 17->21
Noticed a -Wformat-security issue when testing linux-headers reverse
dependencies and ended up spending far too much time on hunting down
the last version of this in Debian.

Thanks to Kangie for digging up the needed files! It turns out that
the Debian patches tarball in this final release (22 isn't useful as it's
a transition package, so 21 is the last one) contains *solely* this
-Wformat-security fix, but oh well.

While here, fix a <stdint.h> issue for musl too.

Closes: https://bugs.gentoo.org/713714
Thanks-to: Matt Jolly <kangie@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>