wine.eclass: do not pass CFLAGS as CROSSCFLAGS for mingw by default
Desktop integration tab wrt bug #960825 is not *that* big of an issue,
but these kind of issues with CROSSCFLAGS="-march=native ..." keep
coming back and I'd rather this not be done anymore unless
USE=custom-cflags so regular users do not have to deal with this.
Very few people/distros set CROSSCFLAGS and by default wine will use
just -O2 rather than try to copy CFLAGS which results in a lot of
(not really) "Gentoo"-specific obscure issues.

Closing bug given not currently planning to change this even if that
specific issue is fixed in upstream gcc/wine, but feel free to
re-open/repurpose if someone wants to pursue this further.

Technically needs revbumps, but letting this propagate on wine bumps
should be sufficient, wine should still mostly work albeit the exact
extend of issues beyond the desktop integration tab is unknown. Please
rebuild manually if affected and want this fixed now.

Leaving this alone for clang (USE=-mingw), no issues that I'm aware
of yet (or at least not specific to *FLAGS, it is generally less
tested with wine and may have other unrelated issues).

Closes: https://bugs.gentoo.org/960825
Thanks-to: GoldsrcSource
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>