sci-libs/clblas: add trivial build dep
To build, this needs an opencl runtime and also an opencl buildtime. :)
opencl provides these in different packages, so it's not sufficient to
depend on the former alone. Although the headers are also a build dep
for the runtime, they might be depcleaned and are never installed at all
when binpkgs are used.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>