dev-python/pytest-twisted: Restore pytest entry point
Restore the pytest entry point, as it is required for EPYTEST_PLUGINS
to work.  Besides, it is against Gentoo policy to aggressively strip
pytest entry points from packages, making their behavior incompatible
to upstream.  We can disable it via the eclass when we have reports
of failures (that is, if people still need plugin autoloading).

Signed-off-by: Michał Górny <mgorny@gentoo.org>