app-text/gtranslator: Needs libgda5 with http USE disabled
This was probably overlooked in the past due to not having "http" USE enabled
for libgda5. If it is enabled, it will load libsoup2.4 symbols and the app
will die with:
libsoup3 symbols detected. Using libsoup2 and libsoup3 in the same process is
not supported

We need then to ensure that libgda5 does not load libsoup2 symbols and,
probably, try to finally package libgda6 (even if I see that upstream did not
tagged any newer version).

Signed-off-by: Pacho Ramos <pacho@gentoo.org>