net-libs/gssdp: remove dependency on old slot
When net-libs/gssdp:1.6 was added in commit 9b9c244c7c5611c550ff38dfdc748dcbf13ab8b6
to the Gentoo tree, depended on previous slot 0 to manage a file collision.
But the older slot still depends on deprecated net-libs/libsoup:2.4.

Now, we switch the logic to prefer the file from the newer slot, and rename
(not delete) the file in the older slot.

Signed-off-by: Cristian Othón Martínez Vera <cfuga@cfuga.mx>