sys-libs/glibc: backport various ebuild fixes to <=2.36-r8
This commit applies several quality improvements to the glibc-2.36-r8
ebuild and its predecessors. These improvements are described herewith.

Refrain from specifying the --inplace-glibc option to locale-gen(1),
fixing builds for which the "compile-locales" USE flag is in effect.

As regards the pkg_postinst() function, raise a warning if locale-gen(1)
does not exit successfully.

As regards the glibc_do_src_install() function, die if locale-gen(1)
does not exit successfully.

Refrain from specifying the --jobs option to locale-gen(1) in the case
that the value of the 'EMERGE_FROM' variable is "binary".

As regards the pkg_postinst() function of the glibc-2.19-r3 ebuild,
check whether the value of the 'ROOT' variable is empty rather than
non-empty, just as all of the other ebuilds do.

See-also: e133de5ee597abf391f5ea3d13e1d077a40a2165
Closes: https://bugs.gentoo.org/915629
Bug: https://bugs.gentoo.org/736794
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>