dev-qt/qtlanguageserver: sync stabilize 6.9.1 for arm64
Not tested in stable, but was barely just tested for keywording
in bug #960936 and stable has the same Qt version making this
very similar and there's no reason not to keep this in sync with
rest of Qt.

Doing this is handy to simplify updating qmlls masks, and stable
use.masks would be annoying/confusing for users accept keywording
(soon to be keyworded) ~arm64 freecad.

Bug: https://bugs.gentoo.org/960937
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>