www-client/firefox: undo --disable-jit on riscv
This builds now. Whether or not the JIT is actually enabled upstream is
another story, but fewer special cases in the ebuild are better. This
should get a revision bump but since it only touches riscv and since
firefox is one of the hardest packages to build, I've left it
alone. You'll get your JIT with v141.0.1.

Bug: https://bugs.gentoo.org/947338
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>