dev-lang/python-exec: add comment explaining PYTHON_COMPAT is cosmetic here
This is just a stub for tooling, see the IUSE hack below. Don't panic
if the array here doesn't contain the latest impl. If you can't use a
new impl., it's likely because it's a stable-masked flag.

Prompted by some confusion on IRC. I actually thought we had a comment
here already.

Signed-off-by: Sam James <sam@gentoo.org>