sys-devel/binutils: drop LTO strip patch from 2.44
This needed some followups in the end so let's back it out of 2.44 and
instead handle it in 2.45 (where most of it is upstream but some fixes
need to be backported).

In particular, see https://sourceware.org/PR33230 and https://sourceware.org/PR33246
which got reported post-release. This may have caused bug #960493 too.

Having 2.44 be "vanilla" in this regard will make it much easier to test
and debug.

TL;DR: Drop the backports we did, the functionality we needed from them
is in 2.45 anyway.

Bug: https://bugs.gentoo.org/960493
Signed-off-by: Sam James <sam@gentoo.org>