gnome-extra/cjs: add 128.0
There are a couple big changes in packaging in this version.

First and foremost is a massive version bump. This is actually aligned
with spidermonkey -- cinnamon devs rightly note that cjs doesn't really
change outside the spidermonkey version it bridges, so the new
versioning both reflects that and makes it easier to track what it uses.
The minor version will be used for cjs-specific bug fixes.

This makes it easy to have a slot-dep in Gentoo. Nice! Also, that means
we now support spidermonkey 128, up from 115, getting us back onto the
"latest version" train, and as a side effect, supporting newer
rust/llvm/python versions.

USE=cairo support is now also removed, as it is now always enabled in
the upstream build.

Frankly this seems like good sense to me. The only revdep depends on
cjs[cairo], how exactly shall it be disabled? And now it may depend on
cjs[cairo(+)] instead.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>