media-video/ffmpeg: fix build with >=nvidia-cuda-toolkit-13
Not in tree yet, but in preparation.

Tested using -vf scale_npp which seems to work properly for cuda-12
with or without the patch. Trying to use 13 at runtime did not work
properly for me but may just be due to my old card being (now)
unsupported in 13, Negril reports that 13 works fine with the patch
so that's good enough for me.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>