sci-libs/ginkgo: fix missing dep by not building examples
examples are not installed, but add ~50 new C++ build targets. In
1.10.0, they also add a DEPEND on nlohmann_json, which then fails by
falling back to git cloning it if a system copy is not found.

```
CMake Error at /usr/share/cmake/Modules/FetchContent.cmake:1918 (message):
  Build step for nlohmann_json failed: 1
Call Stack (most recent call first):
  /usr/share/cmake/Modules/FetchContent.cmake:1609 (__FetchContent_populateSubbuild)
  /usr/share/cmake/Modules/FetchContent.cmake:2145:EVAL:2 (__FetchContent_doPopulation)
  /usr/share/cmake/Modules/FetchContent.cmake:2145 (cmake_language)
  /usr/share/cmake/Modules/FetchContent.cmake:2384 (__FetchContent_Populate)
  third_party/nlohmann_json/CMakeLists.txt:10 (FetchContent_MakeAvailable)

-- Configuring incomplete, errors occurred!
 * ERROR: sci-libs/ginkgo-1.10.0::gentoo failed (configure phase):
```

But we wouldn't need this at all, if we refrained from building examples
we don't use, not even for tests which are already disabled.

Fixes: 64fb6d4f52ca6c0d086289743b74f3a1183e538d
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>