app-text/barcode: properly install static lib, without mangled headers
In 2010, the package was ported to automake and libtool. The library was
dropped in the port; later, barcode.h was updated to include private
headers, like config.h

In Gentoo, we patched it back to install a library again. Crudely, since
it "had" to install /usr/include/gettext.h as well, which is, erm, kind
of Very Bad(tm). It still required config.h as well, which meant it
worked with its revdep "glabels" but perhaps not much else (or at least,
it would need a GNU autoheader using project). Not even a HAVE_CONFIG_H
guard.

Fix that by patching the header so we can avoid installing incorrect
headers.

Take this opportunity as well to improve libtool use by using
lib_LIBRARIES, i.e. don't require libtool to compile a library that is
never shared, don't install *.la files.

Bug: https://bugs.gentoo.org/758491
Closes: https://bugs.gentoo.org/841869
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>