dev-db/mariadb: add 11.4.8
The patches tarball contains numerous changes compared to previous
11.4.7 version:

- Drop "libmariadb: fix mysql_st API regression" patch. All affected
  versions of packages are no longer in the tree. Related upstream
  issue: https://jira.mariadb.org/browse/MDEV-12950

- Add mariadb-10.6.11-gssapi.patch to the tarball. This patch is closely
  related to "libmariadb: cmake: find GSSAPI via pkg-config" patch and
  they are both stored in the tarball now.

- Drop "Allow packagers to use system thrift library" patch as it was
  integrated to the 11.4.8 release.

- Drop "chore(build) update boost to 1.88" patch as it was also
  integrated to the 11.4.8.release.

The disabled_tests array construction was simplified by using a direct
array declarations instead of multiple append operations. This format
was introduced in 11.8.2 ebuild but I want to use it in other Mariadb
series ebuilds as well.

Numerous tests are no longer disabled as they are not causing issues
anymore. Namely:

  compat/oracle.plugin
  main.func_time
  main.plugin_auth
  mariabackup.*
  perfschema.nesting
  perfschema.prepared_statements
  plugins.auth_ed25519
  plugins.cracklib_password_check
  plugins.two_password_validations
  sys_vars.wsrep_on_without_provider

This version contains the GCC 16 fix proposed in PR #43263 and the
related mariadb-11.4.7-gcc-16.patch is no longer needed as well.

Link: https://github.com/gentoo/gentoo/pull/43263
Signed-off-by: Petr Vaněk <arkamar@gentoo.org>