meson.eclass: fix interaction with multilib.eclass
meson.eclass constructs a Meson "native file" using the build-machine
toolchain configuration derived by "$(tc-getBUILD_CC)" et al. This is
correct for building code to run on the _build_ machine, but it may not
be correct for building code to run on the _host_ machine, even in the
case that ${CHOST} == ${CBUILD}, as the ABI flags in ${CC} and
${BUILD_CC} may differ.

In the case that ${CHOST} == ${CBUILD} but ${ABI} != ${DEFAULT_ABI}, we
can put Meson into cross-build mode but set "needs_exe_wrapper = false"
in the "cross file" so that any build-time feature tests need not fall
back on hard-coded defaults.

Consider what happens with Meson _not_ in cross-build mode:

 * The "native file" contains the build-machine toolchain configuration,
   which always builds for the default ABI.
 * For each targeted host ABI, multilib.eclass appends
   "$(get_abi_CFLAGS)" to the host-machine toolchain variables, but this
   has no effect on the Meson "native file."
 * Thus, Meson builds its targets unconditionally for the default ABI.

Here is a trivial reproducer of the above-described failure scenario:

  # CC='x86_64-pc-linux-gnu-gcc-14' \
    BUILD_CC='x86_64-pc-linux-gnu-gcc-14' \
    USE='abi_x86_32' \
    emerge -1 media-libs/fontconfig

The above command line will build what are supposed to be 32-bit object
files during the abi_x86_32 phase of the multi-build, but because it
fails to pass "-m32" to the compiler driver, it actually produces 64-bit
object files, which then cannot be linked together with the 32-bit
/usr/lib/libfreetype.so. After we make the change given in this commit,
the above build succeeds.

Closes: https://bugs.gentoo.org/960506
Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>