meson.eclass: set "needs_exe_wrapper = false" if possible
When the only reason we are running Meson in cross mode is because we
are building for a non-default ABI of the native platform, we can tell
Meson that no wrapper is necessary when executing host-machine binaries
on the build machine.

Also, the native file name really shouldn't be
"meson.${CBUILD}.${ABI}.ini", as ${ABI} refers to an ABI of CHOST, not
of CBUILD, so just drop the ".${ABI}" from the file name since the
native file will not vary between ABI phases.

Closes: https://bugs.gentoo.org/960506
Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>
Signed-off-by: James Le Cuirot <chewi@gentoo.org>