sys-auth/polkit: drop realpath patches
In 21c9a61e14565aeb14f2a58e366ed81f6144c7db, I backported some fixes
for https://github.com/polkit-org/polkit/issues/194 but this ends up
breaking non-merged-usr systems with certain polkit rules.

While this will be a problem in future, right now, we're the only ones
hitting it as a result of the backport, so let's pick the devil we know
by going back to the previous state.

(Committed as stable given I don't see much value in having this in ~arch
and it's been disruptive for people, keeping -r1 in case people want to
test with that for now.)

Closes: https://bugs.gentoo.org/959152
Signed-off-by: Sam James <sam@gentoo.org>