dev-db/mariadb: add 11.8.3
This version bump reuses 11.4.8 patches tarball. See the previous commit
for the list of changes.

The GCC 16 fix proposed in PR #43263 was integrated to this version and
the related mariadb-11.4.7-gcc-16.patch is no longer needed as well.

All tests which used to fail in network-sandbox because of hostname set
to "localhost" were addressed in upstream, therefore, they are removed
from disabled_tests array.

However, "spider.*" related tests are still disabled, because they were
not fully addressed.

Link: https://github.com/gentoo/gentoo/pull/43263
Signed-off-by: Petr Vaněk <arkamar@gentoo.org>