sec-keys.eclass: make test_network actually be respected
Apparently this PROPERTIES has the insidious effect of requiring
multiple moving parts to actually have the desired effect:

- PROPERTIES="test_network"

  Add network access, always, and run tests without comment.

- PROPERTIES="test_network" RESTRICT="test"

  Don't run tests, but if tests ran, they would have network.

- PROPERTIES="test_network" RESTRICT="test" ALLOW_TEST="network"

  Disable RESTRICT, and run tests.

- PROPERTIES="test_network" RESTRICT="!test? ( test )" ALLOW_TEST="network"

  Run tests without build dependencies installed, if USE="-test" is set.

- eclass:
      PROPERTIES="test_network" RESTRICT="!test? ( test ) test" ALLOW_TEST="network"
  ebuild:
      RESTRICT="test" # tests broken, see bug #xxxxxx

  Run broken tests, which fail. Sorry, you're not allowed to communicate
  the desired intent. Don't use eclasses.

This is a total mess, but apparently it is what we must do anyways in
order to have the eclass behave as intended. So, switch from the first
configuration to the fifth, and let's simply pray nobody ever wants to
restrict tests for *any* reason for the package subset known as
"sec-keys/*" and we can safely co-opt its meaning.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>